requirements.txt
Click==7.0
countryinfo==0.1.0
dnspython==1.16.0
eventlet==0.24.1
first==2.0.2
Flask==1.0.2
future==0.17.1
greenlet==0.4.15
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
monotonic==1.5
numpy==1.16.2
python-engineio==3.4.4
python-socketio==4.0.0
reverse-geocode==1.3
scipy==1.2.1
six==1.12.0
Werkzeug==0.14.1
